Initial splitting of polars utils
